kripto/requirements.txt
Flask
pycryptodome 